+++ ask-smapi-model/ask_smapi_model/v1/skill/interaction_model/slot_definition.py
@@ -35,6 +35,8 @@ class SlotDefinition(object):
     :type name: (optional) str
     :param object_type: The type of the slot. It can be a built-in or custom type.
     :type object_type: (optional) str
+    :param generated_by: Name of the generator used to generate this object.
+    :type generated_by: (optional) str
     :param multiple_values: Configuration object for multiple values capturing behavior for this slot.
     :type multiple_values: (optional) ask_smapi_model.v1.skill.interaction_model.multiple_values_config.MultipleValuesConfig
     :param samples: Phrases the user can speak e.g. to trigger an intent or provide value for a slot elicitation.
@@ -44,6 +46,7 @@ class SlotDefinition(object):
     deserialized_types = {
         'name': 'str',
         'object_type': 'str',
+        'generated_by': 'str',
         'multiple_values': 'ask_smapi_model.v1.skill.interaction_model.multiple_values_config.MultipleValuesConfig',
         'samples': 'list[str]'
     }  # type: Dict
@@ -51,19 +54,22 @@ class SlotDefinition(object):
     attribute_map = {
         'name': 'name',
         'object_type': 'type',
+        'generated_by': 'generatedBy',
         'multiple_values': 'multipleValues',
         'samples': 'samples'
     }  # type: Dict
     supports_multiple_types = False
 
-    def __init__(self, name=None, object_type=None, multiple_values=None, samples=None):
-        # type: (Optional[str], Optional[str], Optional[MultipleValuesConfig_83acc8ba], Optional[List[object]]) -> None
+    def __init__(self, name=None, object_type=None, generated_by=None, multiple_values=None, samples=None):
+        # type: (Optional[str], Optional[str], Optional[str], Optional[MultipleValuesConfig_83acc8ba], Optional[List[object]]) -> None
         """Slot definition.
 
         :param name: The name of the slot.
         :type name: (optional) str
         :param object_type: The type of the slot. It can be a built-in or custom type.
         :type object_type: (optional) str
+        :param generated_by: Name of the generator used to generate this object.
+        :type generated_by: (optional) str
         :param multiple_values: Configuration object for multiple values capturing behavior for this slot.
         :type multiple_values: (optional) ask_smapi_model.v1.skill.interaction_model.multiple_values_config.MultipleValuesConfig
         :param samples: Phrases the user can speak e.g. to trigger an intent or provide value for a slot elicitation.
@@ -73,6 +79,7 @@ class SlotDefinition(object):
 
         self.name = name
         self.object_type = object_type
+        self.generated_by = generated_by
         self.multiple_values = multiple_values
         self.samples = samples