+++ ask-smapi-model/CHANGELOG.rst
@@ -251,3 +251,11 @@ This release contains the following changes :
 This release contains the following changes :
 
 - Updating model definitions
+
+
+1.14.5
+^^^^^^
+
+This release contains the following changes :
+
+- Updating model definitions